requirements.txt
django==3.2
djangorestframework==3.13.1
djangorestframework-simplejwt==5.0.0
pytest==6.2.5
pytest-django==4.5.2
pytest-cov==3.0.0
psycopg2-binary==2.9.3